+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   affect wallet functionality.
 - changed "You can add more tokens from the Manage menu." -> "To add more tokens, tap Manage." and
   added item decorator for divider in TokensFragment.kt
+- Added new Ui for raw metadata dialog
 
 ### Changed